requirements.txt
aiogram>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
tenacity>=8.2.0
aiosqlite>=0.19.0
